server/requirements.txt
click==8.1.4
Flask==2.2.5
Flask-Cors==4.0.0
flask-sqlalchemy==3.0.5
greenlet==2.0.2
imageio==2.31.1
img2texture==1.1.0
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.21.6
Pillow==9.5.0
python-dotenv==0.15.0
SQLAlchemy==2.0.18
typing-extensions==4.7.1
waitress==2.1.2
Werkzeug==2.2.3
zipp==3.15.0
